separate RaftGroupId and CPGroupId structs.
