+++ internal/cp/types/raft_group_id.go
@@ -1,0 +1,8 @@
+package types
+
+import "github.com/hazelcast/hazelcast-go-client/types"
+
+type RaftGroupId struct {
+	*types.CPGroupId
+	Seed int64
+}
