+++ internal/proto/codec/atomic_long_alter_codec.go
@@ -17,9 +17,9 @@
 package codec
 
 import (
+	"github.com/hazelcast/hazelcast-go-client/internal/cp/types"
 	"github.com/hazelcast/hazelcast-go-client/internal/proto"
 	iserialization "github.com/hazelcast/hazelcast-go-client/internal/serialization"
-	"github.com/hazelcast/hazelcast-go-client/types"
 )
 
 const (
