+++ internal/proto/codec/atomic_long_get_and_set_codec.go
@@ -17,8 +17,8 @@
 package codec
 
 import (
+	"github.com/hazelcast/hazelcast-go-client/internal/cp/types"
 	"github.com/hazelcast/hazelcast-go-client/internal/proto"
-	"github.com/hazelcast/hazelcast-go-client/types"
 )
 
 const (
