+++ internal/proto/codec/raft_group_id_codec.go
@@ -17,6 +17,7 @@
 package codec
 
 import (
+	types2 "github.com/hazelcast/hazelcast-go-client/internal/cp/types"
 	"github.com/hazelcast/hazelcast-go-client/internal/proto"
 	"github.com/hazelcast/hazelcast-go-client/types"
 )
@@ -27,7 +28,7 @@ const (
 	RaftGroupIdCodecIdInitialFrameSize = RaftGroupIdCodecIdFieldOffset + proto.LongSizeInBytes
 )
 
-func EncodeRaftGroupId(clientMessage *proto.ClientMessage, raftGroupId types.RaftGroupId) {
+func EncodeRaftGroupId(clientMessage *proto.ClientMessage, raftGroupId types2.RaftGroupId) {
 	clientMessage.AddFrame(proto.BeginFrame.Copy())
 	initialFrame := proto.NewFrame(make([]byte, RaftGroupIdCodecIdInitialFrameSize))
 	FixSizedTypesCodec.EncodeLong(initialFrame.Content, RaftGroupIdCodecSeedFieldOffset, int64(raftGroupId.Seed))
@@ -39,7 +40,7 @@ func EncodeRaftGroupId(clientMessage *proto.ClientMessage, raftGroupId types.Raf
 	clientMessage.AddFrame(proto.EndFrame.Copy())
 }
 
-func DecodeRaftGroupId(frameIterator *proto.ForwardFrameIterator) types.RaftGroupId {
+func DecodeRaftGroupId(frameIterator *proto.ForwardFrameIterator) types2.RaftGroupId {
 	// begin frame
 	frameIterator.Next()
 	initialFrame := frameIterator.Next()
@@ -48,10 +49,10 @@ func DecodeRaftGroupId(frameIterator *proto.ForwardFrameIterator) types.RaftGrou
 
 	name := DecodeString(frameIterator)
 	CodecUtil.FastForwardToEndFrame(frameIterator)
-
-	return types.RaftGroupId{
-		Name: name,
-		Seed: seed,
-		Id:   id,
-	}
+	return types2.RaftGroupId{
+		CPGroupId: &types.CPGroupId{
+			Name: name,
+			Id:   id,
+		},
+		Seed: seed}
 }
